job_server/requirements.txt
boto3==1.3.0
PyYAML==3.11
SQLAlchemy==1.0.12
SQLAlchemy-Utils==0.32.0
tornado==4.3